pythonProject/LangChainAcademy/langchain-academy/template/requirements.txt
langgraph
langchain-openai
langchain-core
python-dotenv
pytest